refacto(scopes): split and clean up Scope module
