--- packages/scopes/src/Ref.ts
@@ -1,7 +1,0 @@
-export type Ref = {}
-
-export const create = (): Ref => ({})
-
-export const Ref = {
-  create
-}
