+++ packages/scopes/src/refs/index.ts
@@ -1,0 +1,13 @@
+export const REF_SYMBOL: unique symbol = Symbol('Ref')
+
+export type Ref = {
+  [REF_SYMBOL]: boolean
+}
+
+export const create = (): Ref => ({
+  [REF_SYMBOL]: true
+})
+
+export const Ref = {
+  create
+}
