+++ packages/scopes/src/resources/index.ts
