+++ packages/scopes/src/scopes/symbols.ts
@@ -1,0 +1,3 @@
+export const SCOPE_SYMBOL: unique symbol = Symbol('Scope')
+export const SCOPE_INTERNAL: unique symbol = Symbol('Scope.Internal')
+export const SCOPE_HIERARCHY: unique symbol = Symbol('Scope.Hierarchy')
