+++ packages/scopes/src/variables/constants.ts
@@ -1,2 +1,14 @@
-export const VAR_SYMBOL: unique symbol = Symbol('Var')
-export const ABSTRACT_SYMBOL: unique symbol = Symbol('Var.Abstract')
+import { pipe, Prom } from '@apoyo/std'
+import { Resource } from '../resources'
+
+import { create, Var } from './core'
+
+export const thunk = <T>(thunk: () => T | PromiseLike<T>): Var<T> =>
+  create(async (ctx) => ({
+    scope: ctx.scope.root,
+    mount: () => pipe(Prom.thunk(thunk), Prom.map(Resource.of))
+  }))
+
+export const of = <T>(value: T): Var<T> => thunk(() => value)
+
+export const empty = of<void>(undefined)
