+++ packages/scopes/src/variables/core.ts
@@ -1,11 +1,9 @@
-import { pipe, Prom, Dict } from '@apoyo/std'
-import { Ref } from '../Ref'
-import { Resource } from '../Resource'
-import { Scope } from '../Scope'
+import { Dict } from '@apoyo/std'
+import { Ref } from '../refs'
+import { Resource } from '../resources'
+import { Scope } from '../scopes'
 import { Context } from '../types'
-import { getRoot } from '../utils'
-import { ABSTRACT_SYMBOL, VAR_SYMBOL } from './constants'
-import { getLowestScope } from './utils'
+import { ABSTRACT_SYMBOL, VAR_SYMBOL } from './symbols'
 
 export interface Var<T = any> {
   [VAR_SYMBOL]: Ref
@@ -53,16 +51,6 @@ export const factory = <T, Fun>(factory: Fun, variable: Var<T>): Var.Factory<T, 
   factory
 })
 
-export const thunk = <T>(thunk: () => T | PromiseLike<T>): Var<T> =>
-  create(async (ctx) => ({
-    scope: getRoot(ctx.scope),
-    mount: () => pipe(Prom.thunk(thunk), Prom.map(Resource.of))
-  }))
-
-export const of = <T>(value: T): Var<T> => thunk(() => value)
-
-export const empty = of<void>(undefined)
-
 export const lazy = <A>(fn: () => PromiseLike<Var<A>> | Var<A>): Var<A> =>
   create(async (ctx) => Promise.resolve().then(fn).then(ctx.scope.load))
 
@@ -77,30 +65,4 @@ export const resource = <A, B>(fn: (value: A) => Resource<B> | PromiseLike<Resou
         mount: () => ctx.scope.get(variable).then(fn)
       }
     })
-  )
-
-export const map = <A, B>(fn: (value: A) => B | PromiseLike<B>) => (variable: Var<A>) =>
-  factory(
-    fn,
-    pipe(
-      variable,
-      resource((v) => Promise.resolve(v).then(fn).then(Resource.of))
-    )
-  )
-
-export const chain = <A, B>(fn: (value: A) => PromiseLike<Var<B>> | Var<B>) => (variable: Var<A>) =>
-  factory(
-    fn,
-    create<B>(
-      async (ctx): Promise<Var.Loader> => {
-        const chainedVar = await ctx.scope.get(variable).then(fn)
-        const createdVar = await ctx.scope.load(chainedVar)
-        const created = await ctx.scope.load(variable)
-
-        return {
-          scope: getLowestScope(ctx.scope, [created.scope, createdVar.scope]),
-          mount: () => ctx.scope.get(chainedVar).then(Resource.of)
-        }
-      }
-    )
   )
