+++ packages/scopes/src/variables/struct.ts
@@ -1,9 +1,9 @@
 import { Dict, pipe, Task } from '@apoyo/std'
 
-import { Resource } from '../Resource'
+import { Resource } from '../resources'
+import { Scope } from '../scopes'
 import { Context } from '../types'
 import { create, Var } from './core'
-import { getLowestScope } from './utils'
 
 export function struct<A extends Dict<Var>>(obj: A): Var.Struct<A>
 export function struct(obj: Dict<Var>): Var<Dict>
@@ -11,7 +11,7 @@ export function struct(obj: Dict<Var>): Var<Dict> {
   return create(
     async (ctx: Context): Promise<Var.Loader> => {
       const created = await pipe(obj, Dict.map(Task.taskify(ctx.scope.load)), Task.struct(Task.all))
-      const scope = getLowestScope(
+      const scope = Scope.getLowestScope(
         ctx.scope,
         pipe(
           created,
