+++ packages/scopes/src/variables/symbols.ts
@@ -1,0 +1,2 @@
+export const VAR_SYMBOL: unique symbol = Symbol('Var')
+export const ABSTRACT_SYMBOL: unique symbol = Symbol('Var.Abstract')
